+++ workspaces/x2a/plugins/x2a-common/src/domain/index.ts
@@ -14,4 +14,6 @@
  * limitations under the License.
  */
 
+export { JobStatus } from './JobStatus';
 export { Phase } from './Phase';
+export { ProjectState } from './ProjectState';